+++ playground/src/main.tsx
@@ -4,7 +4,10 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import { theme } from './components/theme/theme.ts';
+
 import './index.css';
+import './components/css/plyr.css';
+import './components/css/styles.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>